feat: create a router component to separates routes from App file
