+++ src/App.tsx
@@ -1,25 +1,13 @@
 import * as React from 'react';
-import {createBrowserRouter, RouterProvider} from 'react-router-dom';
-import TeamOverview from './pages/TeamOverview';
-import Teams from './pages/Teams';
-import UserOverview from './pages/UserOverview';
+import {ThemeProvider} from 'styled-components';
 
-const App = () => {
-    const router = createBrowserRouter([
-        {
-            path: '/',
-            element: <Teams />,
-        },
-        {
-            path: '/team/:teamId',
-            element: <TeamOverview />,
-        },
-        {
-            path: '/user/:useId',
-            element: <UserOverview />,
-        },
-    ]);
-    return <RouterProvider router={router} />;
-};
+import {defaultTheme} from 'styles/themes/default';
+import Router from 'Router';
+
+const App = () => (
+    <ThemeProvider theme={defaultTheme}>
+        <Router />
+    </ThemeProvider>
+);
 
 export default App;
